refactor: Pitcher 숫자 검증로직 분리
